conf: Add flag for whether the rotation was configured or not.
If the user specified a rotation the value could be 0,90,180 or 270.
If they didn't specify anything it will be 0. Currently it's not
possible to tell if they asked for 0 or it's just the default.

If the value is 0 but the user didn't set it we should ask the
system what it thinks the rotation is (i.e. the panel orientation
from DRM/KMS). So a flag is needed to tell if the 0 came from the
user.

Add a flag for whether the rotation was user set.